move `_MemberSpec` into `_memberspec` module
